test(core): Functional model tests to ensure linking models to ticket via ticket comment creates an action comment
ref: #1063 #1062

+++ app/core/tests/functional/slash_commands/test_linked_model.py
@@ -1,10 +1,17 @@
 import pytest
 
+from django.apps import apps
+
 from core.tests.functional.slash_commands.test_slash_command_related import SlashCommandsCommon
 
 
 
 class LinkedModelFixtures:
+    """Model Specific ticket tests
+
+    This test suite is intended to be included within model Functional test
+    suites.
+    """
 
 
 
@@ -48,7 +55,7 @@ class LinkedModelFixtures:
 
         yield ticket_comment
 
-        if getattr(created_model, '_ticket_linkable', False):
+        if getattr(created_model, '_ticket_linkable', False) and ticket_comment.pk:
             ticket_comment.delete()
 
 
@@ -133,11 +140,12 @@ class LinkedModelTicketCommentTestCases(
 
 
     @pytest.mark.module_core
-    @pytest.mark.model_ticketbase
+    @pytest.mark.model_ticketcommentbase
+    @pytest.mark.slash_command
+    @pytest.mark.slash_command_linked_model
     @pytest.mark.tickets
-    @pytest.mark.tickets_slash_command
-    def test_slash_command_link_ticket_comment(self, 
-        created_model, ticket_comment,
+    def test_slash_command_link_ticket_comment(self, mocker,
+        created_model, ticket_comment, model_ticketcommentbase,
         parameterized, param_key_slash_command,
         param_link, param_slash_command,
         param_text, param_stays_in_comment,
@@ -163,6 +171,12 @@ class LinkedModelTicketCommentTestCases(
             )
         )
 
+        context = mocker.patch('core.mixins.centurion.Centurion.context', {
+            'logger': None,
+            model_ticketcommentbase._meta.model_name: ticket_comment.user.user,
+            created_model._meta.model_name: ticket_comment.user.user,
+        })
+
 
         ticket_comment.save()
 
@@ -176,13 +190,15 @@ class LinkedModelTicketCommentTestCases(
 
     @pytest.mark.module_core
     @pytest.mark.model_ticketbase
+    @pytest.mark.slash_command
+    @pytest.mark.slash_command_linked_model
     @pytest.mark.tickets
-    @pytest.mark.tickets_slash_command
-    def test_slash_command_link_ticket(self,
+    def test_slash_command_link_ticket(self, mocker,
         created_model, ticket,
         parameterized, param_key_slash_command, param_name,
         param_link, param_slash_command,
         param_text, param_stays_in_comment,
+        model_ticketbase,
     ):
         """Slash command Check
 
@@ -205,6 +221,12 @@ class LinkedModelTicketCommentTestCases(
             )
         )
 
+        context = mocker.patch('core.mixins.centurion.Centurion.context', {
+            'logger': None,
+            model_ticketbase._meta.model_name: ticket.opened_by.user,
+            created_model._meta.model_name: ticket.opened_by.user,
+        })
+
         ticket.save()
 
 
@@ -212,6 +234,63 @@ class LinkedModelTicketCommentTestCases(
             (f'/{param_slash_command}' in ticket.description) == param_stays_in_comment
             and (command_obj in ticket.description) == param_stays_in_comment
         )
+
+
+
+    @pytest.mark.module_core
+    @pytest.mark.model_ticketcommentaction
+    @pytest.mark.slash_command
+    @pytest.mark.slash_command_linked_model
+    @pytest.mark.tickets
+    def test_slash_command_link_ticket_comment_creates_action_comment(self, 
+        mocker, model,
+        created_model, ticket_comment,
+        parameterized, param_key_slash_command,
+        param_link, param_slash_command,
+        param_text, param_stays_in_comment,
+        model_ticketcommentaction, model_ticketcommentbase
+    ):
+        """Slash command Check
+
+        Ensure that an action comment is created and linked to the ticket.
+        """
+
+        if not getattr(created_model, '_ticket_linkable', False):
+            pytest.xfail( reason = 'Model is not ticket linkable. Test is N/A.' )
+
+
+        if param_stays_in_comment:
+            pytest.xfail( reason = 'slash command is invalid. Test is N/A.' )
+
+
+        context = mocker.patch('core.mixins.centurion.Centurion.context', {
+            'logger': None,
+            model_ticketcommentbase._meta.model_name: ticket_comment.user.user,
+        })
+
+        comment_text = param_text
+
+        assert 'COMMAND' in comment_text
+        # COMMAND must be in ticket comment so it can be constructed
+
+        command_obj = str( f'${created_model.model_tag}-{created_model.id}' )
+
+        ticket_comment.body = str(
+            comment_text.replace(
+                'COMMAND', f'/{param_slash_command} ' + command_obj
+            )
+        )
+
+
+        ticket_comment.save()
+
+        action_comment = model_ticketcommentaction.objects.filter(
+            ticket = ticket_comment.ticket,
+            body = f'linked model {command_obj}'
+        )
+
+
+        assert len(action_comment) == 1
